pyproject.toml
[tool.poetry]
name = "isamples_inabox"
version = "0.5.1"
description = "Python implementation of iSamples-in-a-box components"
authors = ["datadavev <605409+datadavev@users.noreply.github.com>"]
license = "Apache 2.0"
packages = [
    { include = "isb_lib"},
    { include = "isb_web"},
    { include = "isamples_metadata"},
]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
click-config-file = "^0.6.0"
fastapi = "0.104.0"
uvicorn = {version = "^0.23.2", extras=["standard"]}
aiofiles = "^23.2.1"
igsn-lib = {git = "https://github.com/isamplesorg/igsn_lib.git", rev = "main"}
isamples_frictionless = {git = "https://github.com/isamplesorg/isamples_frictionless.git", rev = "main"}
term_store = {git = "https://github.com/isamplesorg/term_store.git", rev = "main"}
accept-types = "^0.4.1"
shapely = "^2.0.2"
geojson = "^2.5.0"
Jinja2 = "^3.1.2"
aiofile = "^3.8.8"

sqlmodel = "^0.0.12"
Authlib = "^1.2.1"
itsdangerous = "^2.1.2"
starlette-oauth2-api = "^0.2.6"
httpx = "^0.25.0"

regex = "^2023.10.3"
h3 = {version = "^4.0.0b2", allow-prereleases = true}
aiodns = "^3.1.1"
aiohttp = "^3.8.6"
rdflib = "^7.0.0"

connegp = "^0.2"
openpyxl = "^3.1.2"
xlrd = "^2.0.1"

PyJWT = "^2.8.0"
petl = "^1.7.14"
ijson = "^3.2.3"

[tool.poetry.dev-dependencies]
pytest = "*"
jupyter = "*"
nbterm = "*"
flake8 = "*"
pipdeptree = "*"
coverage = "*"
pytest-cov = "*"
setuptools = "*"

[tool.poetry.scripts]
sesar_things = "scripts.sesar_things:main"
geome_things = "scripts.geome_things:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
